+++ assets/jsx/workflow-editor/components/keyboard-shortcuts/constants.jsx
@@ -11,3 +11,5 @@ export const SHORTCUT_TOGGLE_DEVELOPER_MODE =
     "publishpress-future/workflow-editor/toggle-developer-mode";
 export const SHORTCUT_TOGGLE_ADVANCED_SETTINGS =
     "publishpress-future/workflow-editor/toggle-advanced-settings";
+export const SHORTCUT_SAVE_AS_CURRENT_STATUS =
+    "publishpress-future/workflow-editor/save-as-current-status";